Use single custom datetime definition
