+++ microservices/nris_api/backend/app/nris/custom_types/__init__.py
@@ -1,10 +1,1 @@
-from datetime import datetime
-from app.extensions import db
 
-class MyDateTime(db.TypeDecorator):
-    impl = db.DateTime
-
-    def process_bind_param(self, value, dialect):
-        if type(value) is str:
-            return datetime.strptime(value[:10], '%Y-%m-%d')
-        return value
