+++ microservices/nris_api/backend/app/nris/models/inspection.py
@@ -5,17 +5,10 @@ from sqlalchemy.orm import validates
 from sqlalchemy.ext.associationproxy import association_proxy
 from app.nris.utils.base_model import Base
 
+from app.nris.custom_types.sl_date_time import MyDateTime
 from app.nris.models.inspection_status import InspectionStatus
 from app.nris.models.inspected_location import INSPECTED_LOCATION_RESPONSE_MODEL
 from app.nris.models.document import DOCUMENT_RESPONSE_MODEL
-
-class MyDateTime(db.TypeDecorator):
-    impl = db.DateTime
-
-    def process_bind_param(self, value, dialect):
-        if type(value) is str:
-            return datetime.strptime(value[:10], '%Y-%m-%d')
-        return value
 
 INSPECTION_RESPONSE_MODEL = api.model(
     'inspection', {
