+++ microservices/nris_api/backend/app/nris/models/noncompliance_legislation.py
@@ -5,14 +5,7 @@ from sqlalchemy.orm import validates
 from sqlalchemy.ext.associationproxy import association_proxy
 from app.nris.utils.base_model import Base
 
-class MyDateTime(db.TypeDecorator):
-    impl = db.DateTime
-
-    def process_bind_param(self, value, dialect):
-        if type(value) is str:
-            return datetime.strptime(value[:10], '%Y-%m-%d')
-        return value
-
+from app.nris.custom_types.sl_date_time import MyDateTime
 
 NONCOMPLIANCE_LEGISLATION_RESPONSE_MODEL = api.model(
     'noncompliance_legislation', {
